# openapi-schema.yaml
openapi: 3.0.0
info:
  title: NaviDocs API
  description: Comprehensive boat documentation management API with 16 supporting tables
  version: 1.0.0
  contact:
    name: NaviDocs Team
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: https://api.navidocs.com/v1
    description: Production server
paths:
...
  /boats/{boatId}/maintenance:
    get:
      summary: List maintenance records for a boat
      operationId: listMaintenanceRecords
      parameters:
        - name: boatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MaintenanceRecord'
    post:
      summary: Create a new maintenance record
      operationId: createMaintenanceRecord
      parameters:
        - name: boatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMaintenanceRecordRequest'
      responses:
        '201':
          description: Record created successfully
  /maintenance/{recordId}:
    get:
      summary: Get a specific maintenance record
      operationId: getMaintenanceRecord
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRecord'
    put:
      summary: Update a maintenance record
      operationId: updateMaintenanceRecord
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMaintenanceRecordRequest'
      responses:
        '200':
          description: Record updated successfully
    delete:
      summary: Delete a maintenance record
      operationId: deleteMaintenanceRecord
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Record deleted successfully

  # Camera Feeds Endpoints
...
components:
  schemas:
...
    MaintenanceRecord:
      type: object
      properties:
        id:
          type: integer
        boatId:
          type: integer
        serviceType:
          type: string
        date:
          type: string
          format: date
        provider:
          type: string
        cost:
          type: number
          format: decimal
        nextDueDate:
          type: string
          format: date
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateMaintenanceRecordRequest:
      type: object
      properties:
        serviceType:
          type: string
        date:
          type: string
          format: date
        provider:
          type: string
        cost:
          type: number
          format: decimal
        nextDueDate:
          type: string
          format: date
        notes:
          type: string

    # Camera Feed Schema
...